### 2018-02-26 #198 @raymondben
antanym
Provides access to Antarctic geographic place name information, and tools for working with those names.

```
Package: antanym
Title: Antarctic Geographic Place Names
Version: 0.3.0
Authors@R: c(person("Ben", "Raymond", email = "ben.raymond@aad.gov.au", role = c("aut", "cre")),
	     person("Michael", "Sumner", email = "michael.sumner@aad.gov.au", role="aut"),
	     person("Ursula", "Harris", role = "ctb"),
	     person("Andrew", "Cowie", role = "ctb"),
	     person("Fraser", "Morgan", role = "ctb")
	    )
Description: Antarctic geographic names from the SCAR Composite Gazetteer of Antarctica, and
  functions for working with those place names.
  assertthat,
  C50,
  geosphere,
  httr,
  raster,
  sp,
  stringi
  covr,
  leaflet,
  rgdal,
  rworldmap
X-schema.org-applicationCategory: Antarctic/Southern Ocean
X-schema.org-keywords: Antarctic, Southern Ocean, place names, gazetteer
X-schema.org-isPartOf: https://scar.org

```

https://github.com/SCAR/antanym

Data retrieval (it retrieves the SCAR Composite Gazetteer data from its host server), and geospatial data (geographic place names are obviously geospatial in nature).

Antarctic researchers, particularly those wanting to produce maps or figures of Antarctic regions. Such figures often need spatial features to be labelled (ice shelves, glaciers, stations, etc). The gazetteer data can also be useful in a quantitative sense, for example for calculating distance to the nearest station (which could perhaps be used as a proxy for human disturbance in an ecological modelling scenario).

The `geonames` package provides access to the global geonames.org database of place names, via their API. The geonames.org database includes place name information from the SCAR Composite Gazetteer of Antarctica (CGA). However, there seems to be a good case for a separate package that deals specifically with Antarctic place names:

1. the SCAR CGA is the authoritative source for Antarctic place name information, and we (SCAR) are the custodians of it. geonames.org has ingested CGA information into their global database and is thus a secondary source of this information. At the time of writing the geonames.org copy of the CGA is out of date (e.g. it does not contain Lassesen Island, Ginger Reef, or Pavlova Island, all added in 2017).

1. antanym provides more information about features than does geonames, including the narrative (description of how a feature came to be given its name and who or what it was named after), the date on which it was named, information about the source of the name, and more.

1. geonames requires a login because it consumes the geonames.org API; antanym does not.

1. much of the geonames package functionality is irrelevant in an Antarctic context (e.g. `GNcities()`, `GNcountry*()`, and `GNfindNearbyPostalCodes()` - there are no cities, countries, or post codes in Antarctica).

1. conversely, antanym has functionality specific to the CGA that geonames does not. In particular, the CGA is a composite gazetteer, which means that a single feature can have multiple names (given by different naming authorities from different countries). Antanym has functions to help resolve these, and to discover all of the names associated with a feature. Using the geonames package, it does not appear to be directly possible to find all of the names associated with a given feature. (As an aside, the geonames.org database must hold this multiple-names information in some form, because using the geonames.org `get` API method (http://api.geonames.org/get?geonameId=6627943&username=demo) shows the alternate names for Booth Island. However, unless I've missed it, the geonames package doesn't wrap the `get` API method.)
Antanym also has functions for suggesting names to add to a map and thinning a list of names to get a more visually pleasing spatial coverage for plotting. Because of the differences in functionality, we have not made any particular attempt to align antanym's interface with that of the geonames package. Antanym's interface is instead loosely modelled on a dplyr-style approach to filtering and subsetting.

1. antanym deals only with the CGA, so it's small enough to cache locally and use offline, whereas geonames is a set of wrappers around the geonames.org API and can't be used offline (though you could `memoise` particular calls and cache those, but then you couldn't issue new queries while offline). This might sound like a trivial issue but is a genuine consideration for Antarctic workers with limited internet access in the field, on station, or at sea.

1. as time permits, we will add other Antarctic-related gazetteers to antanym, which are not available through geonames. These include subantarctic gazetteers and informal gazetteers managed by SCAR and the Australian Antarctic Data Centre.

Clean bill of health!

Barry Rowlingson (github barryrowlingson) - author of the geonames package and all around spatial wizard.
