### 2018-02-28 b43a6ad
check class of input gaz parm
