+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: antanym
 Type: Package
 Title: Antarctic Geographic Place Names
-Version: 0.3.0
+Version: 0.3.1
 Authors@R: c(person("Ben", "Raymond", email = "ben.raymond@aad.gov.au", role = c("aut", "cre")),
 	     person("Michael", "Sumner", email = "michael.sumner@aad.gov.au", role="aut"),
 	     person("Ursula", "Harris", role = "ctb"),
