### 2018-03-08 2784c31
revise column descriptions, drop some non-useful cols
