### 2018-03-20 #198 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @raymondben    !!

- You can put a ropensci review badge in your README

- Goodpractice output: really only the long lines are something to change at some point

```r
It is good practice to

    R/load.R:71:NA
    R/load.R:72:NA
    R/load.R:73:NA
    R/load.R:74:NA
    R/load.R:97:NA
    ... and 6 more lines

    R/load.R:59:1
    R/load.R:64:1
    R/load.R:67:1
    R/load.R:71:1
    R/load.R:79:1
    ... and 71 more lines
```

Seeking reviewers now 🕐

---

Reviewers:  xx xx
Due date: xxx
