### 2018-05-02 #198 @johnbaums
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Thanks for the opportunity to review `antanym`. I see clear value in the package; I have experience working with datasets compiled from multiple sources, and functions that permit sensible and straightforward filtering and subsetting are a lifesaver. I was very impressed by the creative approach to thinning place names (`an_thin`), using experts' perceptions of importance, for printing on maps - bravo! I believe `antanym` will be a great contribution to the R community.

I appreciate the effort the authors have made in ensuring the code is clear, concise, readable, and well documented. The package is well structured, with individual functions that perform clear, separate tasks, and internal helpers that prevent clutter and repetition. Data caching is a great convenience - the authors raise an important point about the relevance of this for Antarctic field workers.

The package is very sound from a technical perspective, and I could not find any glaring omissions regarding functionality. I provide a few minor comments below, primarily relating to the docs.

---
**README.Rmd**
- expand SCAR (Scientific Committee on Antarctic Research) on first reference (or use <abbr> tag).
- Fix typo: "gazetter" -> "gazetteer"

**Antanym.Rmd**
- Fix typo: "gazetter" -> "gazetteer"

**DESCRIPTION**
- expand "SCAR" to its full name in the Description field of DESCRIPTION

---
**utils.R**
- "Antarctic" -> "Antarctica" (line 3, utils.R)

**an_preferred()**
- The docs for the argument  `origin_country` state "a place name given by another country will be chosen" - chosen how?
- A couple of object names are a little confusing here: `out <- rbind(in_coi, out_coi)`,  `out` refers to output data, while `out_coi` refers to rows (IDs) that were not included (i.e., were "out")

**an_thin()**
- `this.dist <- as.matrix(dist(as.data.frame(gaz)[, c("longitude", "latitude")]))` can be replaced with `as.matrix(dist(coordinates(gaz)))`
- Add to the docs that `NULL` can be passed to `score_col` if you just want uniform thinning without scores (or alternatively, I guess, that `score_weighting` can be set to 0).
- this could use `sp::spDists` (or the Vincenty function used elsewhere in `antanym`?) to calculate great circle distances if eventual maps are going to be plotted in projected systems. Otherwise clumping might be expected toward the pole.
- I believe you can replace:
  ```R
  if (inherits(gaz, "SpatialPointsDataFrame")) {
      sc <- as.data.frame(gaz)[, score_col]
      sc <- gaz[, score_col]
  }
  if (inherits(sc, "data.frame")) sc <- unlist(sc)
  ```
  with the more efficient
  ```R
  sc <- gaz[[score_col]]
  ```
  ​
**an_suggest()**
- Remove or clarify the long comment (below). It currently seems to contradict the comment within the subsequent`if` statement `if (map_scale>=10e6) {`
```R
## scale >= 10e6: we have full coverage (nearly so for 12mill) of all
## scar_common_ids, so use per-feature predictions for scale < 10e6: use
## predictions by feature properties (except maybe if area of interest lies
## within a catalogued map) stations as special case?
```

**an_filter()**
- the `query` argument is a regular expression, and is split by the line `sterms <- strsplit(query, "[ ,]+")[[1]]`, so I guess place names are expected to be comma- or space-separated. The documentation for that argument should clarify this expectation; rather than stating that it's a regular expression, I would explicitly state that place names should be a character string comma- or space-separated place names or a regular expression. (Of course regex also works, e.g. `query='Place A|Place B|Some other place'`.)
- In relation to the above point, would it make sense to also allow a character vector (length > 1) of place names?
- add `ignore.case=TRUE` to `grepl` calls for `feature_type_name`, `country_name`, `gazetteer` and `cga_source_gazetteer` for consistency with the function's description ("All text-related matches are treated as regular expressions and are case-insensitive.").
- To clarify the relevance of `an_countries(gaz)`, `an_feature_types(gaz)` and `an_cga_sources(gaz)` to `an_filter`, it would be good to include use of these functions in the examples.

​**an_read()**
- To account for paths whose parent does not exist, I'd suggest replacing "The cache directory will be created if it does not exist." with "An attempt will be made to create the cache directory if it does not exist." (or use `recursive = TRUE`, though I don't like functions using that)

---
**General issues**
- The reference link http://www.scar.org/data-products/cga (provided in several of the functions) is dead (website throws 404 Category not found)
- In the examples for several functions, it's probably better to use `tempdir()` rather than `g <- an_read(cache_directory="c:/temp/gaz")`
- consider including argument names in examples, e.g. include `loc=` and `max_distance=` in `nms <- an_near(an_filter(g, feature_type = "Island"), c(100, -66), 20)`. This is probably more important for the vignette, but should also be considered for the examples provided in the function documentation.

---
**Vignette**
I'd like to see a more detailed vignette that includes a bit more explanatory text about the gaz data structure, and includes examples of all of the exported functions (including those documented as secondary functions, such as `an_countries etc.`) provided by the package (there aren't all that many, so it shouldn't be too onerous).

---
Thanks again for the opportunity to review what I think is a valuable and neatly coded contribution. I'm looking forward to seeing it on CRAN!
