### 2018-06-15 1ddaacd
make sure g object is available for tests
