+++ tests/testthat.R
@@ -2,5 +2,4 @@ library(testthat)
 library(magrittr)
 library(antanym)
 
-g <- an_read()
 test_check("antanym")
