+++ tests/testthat/test_utils.R
@@ -1,6 +1,6 @@
 context("tests of general antanym utility functions")
 
 test_that("list of gazetters is as expected",{
-    expect_identical(an_gazetteers(),c("CGA"))
+    expect_identical(an_gazetteers(), c("CGA"))
 })
 
