+++ tests/testthat/test-sp.R
@@ -1,40 +1,41 @@
 context("antanym spatial objects")
 
-gsp <- an_read(sp=TRUE)
-testpt <- c(110,-66)
-namelist <- function(z)sort(z$place_name)
+g <- an_read(cache_directory = test_cache_dir)
+gsp <- an_read(cache_directory = test_cache_dir, sp = TRUE)
+testpt <- c(110, -66)
+namelist <- function(z) sort(z$place_name)
 
 test_that("sp option returns an sp object", {
-    expect_s4_class(gsp,"SpatialPointsDataFrame")
+    expect_s4_class(gsp, "SpatialPointsDataFrame")
 })
 
 test_that("sp versions of functions match non-sp for lon-lat data", {
-    ext <- c(100,110,-70,-60)
-    expect_gt(nrow(an_filter(g,extent=ext)),100)
-    expect_identical(namelist(an_filter(g,extent=ext)),namelist(an_filter(gsp,extent=ext)))
-    expect_identical(an_countries(g),an_countries(gsp))
-    expect_identical(an_feature_types(g),an_feature_types(gsp))
-    expect_identical(an_cga_sources(g),an_cga_sources(gsp))
+    ext <- c(100, 110, -70, -60)
+    expect_gt(nrow(an_filter(g, extent = ext)), 100)
+    expect_identical(namelist(an_filter(g, extent = ext)), namelist(an_filter(gsp, extent = ext)))
+    expect_identical(an_countries(g), an_countries(gsp))
+    expect_identical(an_feature_types(g), an_feature_types(gsp))
+    expect_identical(an_cga_sources(g), an_cga_sources(gsp))
 
     pt <- testpt
-    expect_identical(namelist(an_near(g,pt,50)),namelist(an_near(gsp,pt,50)))
-    pt <- SpatialPoints(cbind(testpt[1],testpt[2]))
+    expect_identical(namelist(an_near(g, pt, 50)), namelist(an_near(gsp, pt, 50)))
+    pt <- SpatialPoints(cbind(testpt[1], testpt[2]))
     projection(pt) <- "+proj=longlat +datum=WGS84 +ellps=WGS84"
-    expect_identical(namelist(an_near(g,pt,50)),namelist(an_near(gsp,pt,50)))
+    expect_identical(namelist(an_near(g, pt, 50)), namelist(an_near(gsp, pt, 50)))
 
-    expect_identical(namelist(an_preferred(g,"Australia")),
-                     namelist(an_preferred(gsp,"Australia")))
-    expect_identical(namelist(an_preferred(g,c("Australia","Poland"))),
-                     namelist(an_preferred(gsp,c("Australia","Poland"))))
+    expect_identical(namelist(an_preferred(g, "Australia")),
+                     namelist(an_preferred(gsp, "Australia")))
+    expect_identical(namelist(an_preferred(g, c("Australia", "Poland"))),
+                     namelist(an_preferred(gsp, c("Australia", "Poland"))))
 
-    expect_identical(an_url(g[1:10,]),an_url(gsp[1:10,]))
+    expect_identical(an_url(g[1:10, ]), an_url(gsp[1:10, ]))
 
-    ext <- c(60,90,-70,-60)
-    suggested <- an_suggest(g,map_extent=ext,map_dimensions=c(100,100))
-    suggestedsp <- an_suggest(gsp,map_extent=ext,map_dimensions=c(100,100))
+    ext <- c(60, 90, -70, -60)
+    suggested <- an_suggest(g, map_extent = ext, map_dimensions = c(100, 100))
+    suggestedsp <- an_suggest(gsp, map_extent = ext, map_dimensions = c(100, 100))
 
[…]
