+++ tests/testthat/test-general.R
@@ -1,5 +1,7 @@
 context("antanym data structures")
 
+g <- an_read(cache_directory = test_cache_dir)
+
 test_that("gaz_id is present and unique", {
     ## some functions rely on gaz_id, check that it's present and unique
     expect_true("gaz_id" %in% names(g))
