+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: antanym
 Type: Package
 Title: Antarctic Geographic Place Names
-Version: 0.3.1
+Version: 0.3.1.9000
 Authors@R: c(person("Ben", "Raymond", email = "ben.raymond@aad.gov.au", role = c("aut", "cre")),
 	     person("Michael", "Sumner", email = "michael.sumner@aad.gov.au", role="aut"),
 	     person("Ursula", "Harris", role = "ctb"),
@@ -29,7 +29,7 @@ Imports:
 Suggests:
   covr,
   dplyr,
-  testthat,
+  testthat (>= 2.0.0),
   knitr,
   leaflet,
   rgdal,
