### 2018-06-15 f79cf6e
simplify caching interface
