### 2018-06-15 7ccba80
fix creation of session cache dir
