### 2018-06-15 80a8c68
test all 3 caching modes
