### 2018-06-15 1fea7f6
temporarily reinstate cache_directory parm for backwards compatibility
