### 2018-06-15 d24f2ec
retry on download failure
