+++ NAMESPACE
@@ -19,6 +19,7 @@ importFrom(assertthat,is.flag)
 importFrom(assertthat,is.string)
 importFrom(geosphere,distVincentySphere)
 importFrom(httr,GET)
+importFrom(httr,RETRY)
 importFrom(httr,config)
 importFrom(httr,content)
 importFrom(httr,http_error)
