### 2018-06-15 3ce58f8
break an_cache_directory into separate function
