+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export("%>%")
+export(an_cache_directory)
 export(an_cga_sources)
 export(an_countries)
 export(an_feature_types)
