### 2018-06-15 ffb9adc
separate man page for an_gazetteers
