+++ man/an_gazetteers.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gazetteers.R
+\name{an_gazetteers}
+\alias{an_gazetteers}
+\title{The place name gazetteers available}
+\usage{
+an_gazetteers()
+}
+\value{
+character vector
+}
+\description{
+Return a character vector that lists all of the gazetteers available through the antanym package. Currently only one gazetteer is available: the SCAR Composite Gazetteer of Antarctica.
+}
+\examples{
+
+an_gazetteers()
+
+}
+\seealso{
+\code{\link{an_read}}, \code{\link{an_filter}}
+}
