### 2018-06-15 bdb1fba
allow sp or raster objects to be passed as filter extent
