+++ R/antanym.R
@@ -14,7 +14,7 @@
 #' @importFrom rappdirs user_cache_dir
 #' @importFrom raster area cellStats crop extent extent<- projectExtent projection<- raster
 #' @importFrom readr read_csv
-#' @importFrom sp coordinates coordinates<- CRS is.projected spTransform SpatialPoints
+#' @importFrom sp bbox coordinates coordinates<- CRS is.projected spTransform SpatialPoints
 #' @importFrom stats dist na.omit predict
 #' @importFrom stringi stri_trans_general
 #' @importFrom utils download.file write.csv
