+++ NAMESPACE
@@ -39,6 +39,7 @@ importFrom(readr,read_csv)
 importFrom(sp,"coordinates<-")
 importFrom(sp,CRS)
 importFrom(sp,SpatialPoints)
+importFrom(sp,bbox)
 importFrom(sp,coordinates)
 importFrom(sp,is.projected)
 importFrom(sp,spTransform)
