### 2018-06-16 86b905e
modify filtering interface
