+++ vignettes/antanym.Rmd
@@ -166,5 +166,5 @@ See the [antanym-demo](https://github.com/AustralianAntarcticDataCentre/antanym-
 
 ## Future directions
 
-Antanym currently only provides information from the SCAR CGA. This does not cover other features that may be of interest to Antarctic researchers, such as those on subantarctic islands or features that have informal names not registered in the CGA. Antanym may be expanded to cover extra gazetters containing such information, at a later date.
+Antanym currently only provides information from the SCAR CGA. This does not cover other features that may be of interest to Antarctic researchers, such as those on subantarctic islands or features that have informal names not registered in the CGA. Antanym may be expanded to cover extra gazetteers containing such information, at a later date.
 
