### 2018-06-16 cfafa99
replace NA lon/lat on one feature
