### 2018-06-16 747fb31
change an_preferred to use cga_source
