+++ R/utils.R
@@ -1,64 +1,3 @@
-#' Find one name per feature
-#'
-#' The SCAR Composite Gazetteer of Antarctic is a compilation of place names provided by many different countries. The composite nature of the CGA means that there may be multiple names associated with a single feature. The \code{an_preferred} function can be used to resolve a single name per feature.
-#'
-#' @references \url{http://www.scar.org/data-products/cga}
-#' @param gaz data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}
-#' @param origin_country character: vector of country names, in order of preference. If a given feature has been named by one of these countries, this place name will be chosen. If the feature in question has not been given a name by one of these countries, a place name given by another country will be chosen. For valid \code{origin_country} values, see \code{\link{an_countries}}
-#'
-#' @return data.frame of results
-#'
-#' @seealso \code{\link{an_read}}, \code{\link{an_countries}}
-#'
-#' @examples
-#' \dontrun{
-#'  g <- an_read(cache = "session")
-#'
-#'  ## get a single name per feature, preferring the
-#'  ##  Polish name where there is one
-#'
-#'  pnames <- an_preferred(g, "Poland")
-#'
-#'  ## names starting with "Sm", preferring US names then
-#'  ##  Australian ones if available
-#'
-#'  g %>% an_filter("^Sm") %>%
-#'    an_preferred(origin_country = c("United States of America", "Australia"))
-#' }
-#'
-#' @export
-an_preferred <- function(gaz, origin_country) {
-    assert_that(inherits(gaz, c("data.frame", "SpatialPointsDataFrame")))
-    assert_that(is.character(origin_country))
-    is_sp <- inherits(gaz, "SpatialPointsDataFrame")
-    if (is_sp) {
-        ## if sp, work on the @data object
-        gaz_sp <- gaz
-        gaz <- gaz@data
-    }
-
-    ## features that have a name from one of our countries of interest
-    in_ids <- unique(gaz$scar_common_id[gaz$country_name %in% origin_country])
-    in_coi <- gaz[gaz$scar_common_id %in% in_ids, ]
-    ## order within scar_common_id by origin_country (with ordering as per appearance in the origin_country vector)
-    ord <- order(in_coi$scar_common_id, factor(in_coi$country_name, levels=origin_country))
-    in_coi <- in_coi[ord, ]
-    in_coi <- in_coi[!duplicated(in_coi$scar_common_id), ] ## take first entry for each scar_common_id
-
-    out_coi <- gaz[!gaz$scar_common_id %in% in_ids, ]
-    out_coi <- out_coi[!duplicated(out_coi$scar_common_id), ] ## take first entry for each scar_common_id
-
-    out <- rbind(in_coi, out_coi)
-
-    if (is_sp) {
-        ## return the subset of gaz_sp corresponding to the rows we just selected
-        gaz_sp[gaz_sp$gaz_id %in% out$gaz_id, ]
-    } else {
-        out
-    }
-}
-
-
 #' Thin names to give approximately uniform spatial coverage
 #'
 #' The provided data.frame of names will be thinned down to a smaller number of names. The thinning process attempts to select a subset of names that are uniformly spatially distributed, while simultaneously choosing the most important names (according to their relative score in the \code{score_col} column.

[…]
