### 2018-06-16 dbeaf28
doc expanded a bit
