+++ R/cga_preferred_names.R
@@ -1,6 +1,6 @@
 #' Find one name per feature in the SCAR Composite Gazetteer
 #'
-#' The SCAR Composite Gazetteer of Antarctic is a compilation of place names provided by many different countries. The composite nature of the CGA means that there may be multiple names associated with a single feature. The \code{an_preferred} function can be used to resolve a single name per feature.
+#' The SCAR Composite Gazetteer of Antarctic is a compilation of place names provided by many different countries. The composite nature of the CGA means that there may be multiple names associated with a single feature. The \code{an_preferred} function can be used to resolve a single name per feature. Specify one or more \code{cga_source} entries and the input \code{gaz} will be filtered to a single name per feature. For features that have multiple names (i.e. appear in multiple source gazetteers) a single name will be chosen, preferring names from the specified \code{cga_source} entries where possible.
 #'
 #' @references \url{http://www.scar.org/data-products/cga}
 #' @param gaz data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}
