### 2018-06-16 05631ce
internal functions to separate R file
