### 2018-06-16 0eb28b4
remove redundant test file
