### 2018-06-16 33317ee
drop deprecated cache_directory parm, filter/prefer names by 'origin' rather than origin_gazetteer/origin_country
