### 2018-06-16 51872b1
doc updates
