+++ R/gazetteers.R
@@ -2,7 +2,7 @@
 #'
 #' Return a character vector that lists all of the gazetteers present in the \code{gaz} data, or (if \code{gaz} was not provided) all of the gazetteers available through the antanym package. Currently only one gazetteer is available: the SCAR Composite Gazetteer of Antarctica.
 #'
-#' @param gaz data.frame or SpatialPointsDataFrame: (optional) as returned by \code{\link{an_read}}
+#' @param gaz data.frame or SpatialPointsDataFrame: (optional) as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}
 #'
 #' @return character vector. If \code{gaz} was provided, this will be a list of all gazetteers present in \code{gaz}. Otherwise, it will be a list of all gazetteers available through the antanym package
 #'
