+++ man/an_get_url.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/utils.R
-\name{an_url}
-\alias{an_url}
+\name{an_get_url}
+\alias{an_get_url}
 \title{Get links to gazetteer entries}
 \usage{
-an_url(gaz)
+an_get_url(gaz)
 }
 \arguments{
 \item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}}
@@ -18,7 +18,7 @@ Each entry in the SCAR Composite Gazetteer of Antarctica has its own web page. T
 \examples{
 \dontrun{
  g <- an_read(cache = "session")
- my_url <- an_url(an_filter(g, "Ufs Island")[1, ])
+ my_url <- an_get_url(an_filter(g, "Ufs Island")[1, ])
  browseURL(my_url)
 }
 }
