+++ NAMESPACE
@@ -5,6 +5,7 @@ export(an_cache_directory)
 export(an_feature_types)
 export(an_filter)
 export(an_gazetteers)
+export(an_get_url)
 export(an_mapscale)
 export(an_near)
 export(an_origins)
@@ -12,7 +13,6 @@ export(an_preferred)
 export(an_read)
 export(an_suggest)
 export(an_thin)
-export(an_url)
 importFrom(C50,predict.C5.0)
 importFrom(assertthat,assert_that)
 importFrom(assertthat,is.flag)
