### 2018-06-16 039db18
updates to thinning function, add test
