### 2018-06-16 1ea5d20
changed variable names
