### 2018-06-16 b32285e
use coordinates() on sp objects
