+++ man/an_preferred.Rd
@@ -35,7 +35,7 @@ The Composite Gazetteer of Antarctica is a compilation of place names provided b
 
 }
 \references{
-\url{http://www.scar.org/data-products/cga}
+\url{https://www.scar.org/data-products/place-names/}
 }
 \seealso{
 \code{\link{an_read}}, \code{\link{an_origins}}
