### 2018-06-16 d501c50
name args in examples
