+++ R/search.R
@@ -14,16 +14,16 @@
 #'  g <- an_read(cache = "session")
 #'
 #'  ## named features within 10km of 110E, 66S
-#'  an_near(g, c(110, -66), 10)
+#'  an_near(g, loc = c(110, -66), max_distance = 10)
 #'
 #'  ## using pipe operator
-#'  g %>% an_near(c(100, -66), 10)
+#'  g %>% an_near(loc = c(100, -66), max_distance = 10)
 #'
 #'  ## with sp objects
 #'  gsp <- an_read(cache = "session", sp = TRUE)
 #'  loc <- SpatialPoints(matrix(c(110, -66), nrow = 1),
-#'    proj4string=CRS("+proj=longlat +datum=WGS84 +ellps=WGS84"))
-#'  an_near(gsp, loc, 10)
+#'    proj4string = CRS("+proj=longlat +datum=WGS84 +ellps=WGS84"))
+#'  an_near(gsp, loc = loc, max_distance = 10)
 #' }
 #'
 #' @export
@@ -73,33 +73,34 @@ an_near <- function(gaz, loc, max_distance) {
 #'  g <- an_read(cache = "session")
 #'
 #'  ## simple search for any place name containing the word 'William'
-#'  an_filter(g, "William")
+#'  an_filter(g, query = "William")
 #'
 #'  ## only those names originating from Australia or USA
-#'  an_filter(g, "William", origin = "Australia|United States of America")
+#'  an_filter(g, query = "William", origin = "Australia|United States of America")
 #'
 #'  ## this search will return no matches
 #'  ## because the actual place name is 'William Scoresby Archipelago'
-#'  an_filter(g, "William Archipelago")
+#'  an_filter(g, query = "William Archipelago")
 #'
 #'  ## you can split the search terms so that each is matched separately:
-#'  an_filter(g, c("William", "Archipelago"))
+#'  an_filter(g, query = c("William", "Archipelago"))
 #'
 #'  ## or use a regular expression
-#'  an_filter(g, "William .* Archipelago")
+#'  an_filter(g, query = "William .* Archipelago")
 #'
-#'  ## or search by feature type
-#'  an_filter(g, "William", feature_type = "Archipelago")
+#'  ## or refine the search using feature type
+#'  an_filter(g, query = "William", feature_type = "Archipelago")
 #'
 #'  ## for more complex text searching, use regular expressions
 #'  ## e.g. names matching "West" or "East"
-#'  an_filter(g, "West|East")
+#'  an_filter(g, query = "West|East")
 #'
 #'  ## e.g. names starting with "West" or "East"
-#'  an_filter(g, "^(West|East)")
+#'  an_filter(g, query = "^(West|East)")
 #'
 #'  ## e.g. names with "West" or "East" appearing as complete words in the name
-#'  an_filter(g, "\\b(West|East)\\b")
+#'  ## ["\\b" matches a word boundary: see help("regex") ]
+#'  an_filter(g, query = "\\b(West|East)\\b")
 #'
 #'  ## filtering by spatial extent
 #'  nms <- an_filter(g, extent = c(100, 120, -70, -65), origin = "Australia")
@@ -114,14 +115,14 @@ an_near <- function(gaz, loc, max_distance) {
 #'  an_filter(g, extent = bbox(my_sp))
 #'
 #'  ## or using the sp form of the gazetteer data
-#'  gsp <- an_read(cache = "session", sp=TRUE)
+#'  gsp <- an_read(cache = "session", sp = TRUE)
 #'  an_filter(gsp, extent = my_sp)
 #'
 #'  ## using pipe operator
-#'  g %>% an_filter("Ross", feature_type = "Ice shelf|Mountain")
[…]
