### 2018-06-17 999c6bc
add separate function to document the data structure
