+++ R/cga_metadata.R
@@ -1,0 +1,57 @@
+#' Information about the Composite Gazetteer of Antarctica data structure
+#'
+#' The Composite Gazetteer of Antarctica data structure (as returned by \code{\link{an_read}}):
+#'
+#' @references \url{https://www.scar.org/data-products/place-names/}
+#'
+#' @param simplified logical: if TRUE, only describe the simplified set of columns (see the equivalent parameter in \code{\link{an_read}})
+#'
+#' @return a data frame with columns "field" and "description"
+#'
+#' @seealso \code{\link{an_read}}
+#'
+#' @examples
+#'
+#' an_cga_metadata()
+#'
+#' @export
+
+an_cga_metadata <- function(simplified = TRUE) {
+    assert_that(is.flag(simplified), !is.na(simplified))
+
+
+    md <- rbind(list(f = "gaz_id", desc = "The unique identifier of each gazetteer entry. Note that the same feature (e.g. \'Browns Glacier\') might have multiple gazetteer entries, each with their own gaz_id, because the feature has been named multiple times by different naming authorities. The scar_common_id value for these entries will be identical, because scar_common_id identifies the feature itself", simp = TRUE),
+                list(f = "scar_common_id", desc = "The unique identifier (in the Composite Gazetteer of Antarctica) of the feature. A single feature may have multiple names, given by different naming authorities", simp = TRUE),
+                list(f = "place_name", desc = "The name of the feature", simp = TRUE),
+                list(f = "place_name_transliterated", desc = "The name of the feature transliterated to simple ASCII characters (e.g. with diacritical marks removed)", simp = TRUE),
+                list(f = "longitude", desc = "The longitude of the feature (negative values indicate degrees west). Note that many features are not point features (e.g. mountains, lakes), in which case the longitude and latitude values are indicative only, generally of the centroid of the feature", simp = TRUE),
+                list(f = "latitude", desc = "The latitude of the feature (negative values indicate degrees south). Note that many features are not point features (e.g. mountains, lakes), in which case the longitude and latitude values are indicative only, generally of the centroid of the feature", simp = TRUE),
+                list(f = "altitude", desc = "The altitude of the feature, in metres relative to sea level. Negative values indicate features below sea level", simp = TRUE),
+                list(f = "feature_type_name", desc = "The feature type (e.g. \'Archipelago\', \'Channel\', \'Mountain\'). See an_feature_types for a full list", simp = TRUE),
+                list(f = "date_named", desc = "The date on which the feature was named", simp = TRUE),
+                list(f = "narrative", desc = "A text description of the feature; may include a synopsis of the history of its name", simp = TRUE),
+                list(f = "named_for", desc = "The person after whom the feature was named, or other reason for its naming. For historical reasons the distinction between 'narrative' and 'named for' is not always obvious", simp = TRUE),
+                list(f = "origin", desc = "The naming authority that provided the name. This is a country name, or organisation name for names that did not come from a national source", simp = TRUE),
+                list(f = "relic", desc = "If TRUE, this name is associated with a feature that no longer exists (e.g. an ice shelf feature that has disappeared)", simp = TRUE),
+                list(f = "gazetteer", desc = "The gazetteer from which this information came (currently only 'CGA')", simp = TRUE),
+                list(f = "meeting_date", desc = "The date on which the name was formally approved by the associated naming authority. This is not available for many names: see the date_named column", simp = FALSE),
+                list(f = "meeting_paper", desc = "References to papers or documents associated with the naming of the feature", simp = FALSE),
[…]
