+++ man/an_cga_metadata.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cga_metadata.R
+\name{an_cga_metadata}
+\alias{an_cga_metadata}
+\title{Information about the Composite Gazetteer of Antarctica data structure}
+\usage{
+an_cga_metadata(simplified = TRUE)
+}
+\arguments{
+\item{simplified}{logical: if TRUE, only describe the simplified set of columns (see the equivalent parameter in \code{\link{an_read}})}
+}
+\value{
+a data frame with columns "field" and "description"
+}
+\description{
+The Composite Gazetteer of Antarctica data structure (as returned by \code{\link{an_read}}):
+}
+\examples{
+
+an_cga_metadata()
+
+}
+\references{
+\url{https://www.scar.org/data-products/place-names/}
+}
+\seealso{
+\code{\link{an_read}}
+}
