### 2018-06-17 40ae399
reorganize and expand vignette
