+++ vignettes/antanym.Rmd
@@ -9,11 +9,15 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+```{r chunkopts, eval=TRUE, echo=FALSE}
+knitr::opts_chunk$set(eval=TRUE, echo=TRUE, message=FALSE, warning=FALSE, tidy=FALSE, cache=FALSE, include=TRUE, dpi=72, fig.width=9, fig.height=6, fig.align="center", results="markup")
+```
+
 ## Overview
 
 This R package provides easy access to Antarctic geographic place name information, and tools for working with those names.
 
-The authoritative source of place names in Antarctica is the Composite Gazetteer of Antarctica (CGA), which is produced by the Scientific Committee on Antarctic Research (SCAR). The CGA consists of approximately 37,000 names corresponding to 19,000 distinct features. It covers features south of 60 &deg;S, including terrestrial and undersea or under-ice.
+The authoritative source of place names in Antarctica is the Composite Gazetteer of Antarctica (CGA), which is produced by the Scientific Committee on Antarctic Research (SCAR). The CGA consists of approximately 37,000 names corresponding to 19,000 distinct features. It covers features south of 60 &deg;S, including terrestrial and undersea or under-ice features.
 
 There is no single naming authority responsible for place names in Antarctica because it does not fall under the sovereignty of any one nation. In general, individual countries have administrative bodies that are responsible for their national policy on, and authorisation and use of, Antarctic names. The CGA is a compilation of place names that have been submitted by representatives of national names committees from 22 countries.
 
@@ -31,7 +35,7 @@ The SCAR Composite Gazetteer of Antarctica is made available under a CC-BY licen
 ## Installing
 
 
-```{r, eval = FALSE}
+```{r eval = FALSE}
 install.packages("devtools")
 devtools::install_github("SCAR/antanym")
 ```
@@ -40,7 +44,7 @@ devtools::install_github("SCAR/antanym")
 
 Start by fetching the names data from the host server. Here we use a temporary cache so that we can re-load it later in the session without needing to re-download it:
 
-```{r message = FALSE, warning = FALSE}
+```{r}
 library(antanym)
 g <- an_read(cache = "session")
 ```
@@ -51,102 +55,219 @@ If you want to be able to work offline, you can cache the data to a directory th
 g <- an_read(cache = "persistent")
 ```
 
-Or for complete control over the persistent cache, specify the directory:
-
-```{r eval = FALSE}
-g <- an_read(cache = "c:/temp/gaz")
-```
-
 If you prefer working with `sp` objects, then this will return a `SpatialPointsDataFrame`:
 ```{r as_sp, eval = FALSE}
 gsp <- an_read(sp = TRUE, cache = "session")
 ```
 
+### Data summary
+
 How many names do we have in total?
-```{r message = FALSE, warning = FALSE}
+```{r}
 nrow(g)
 ```
 
 Corresponding to how many distinct features?
-```{r message = FALSE, warning = FALSE}
+```{r}
 length(unique(g$scar_common_id))
 ```
 
-Find all names associated with feature 1589 (Booth Island) and the country of origin of each name:
-```{r message = FALSE, warning = FALSE}
+We can get a list of the feature types in the data:
+```{r}
+head(an_feature_types(g), 10)
+```
+
+### Data structure
+
+A description of the data structure can be found in the `an_load` function help, (the same information is available via `an_cga_metadata`):
+```{r echo = FALSE}
+knitr::kable(an_cga_metadata())
+```
+
+### Finding names
+
+The `an_filter` function provides a range of options to find the names you are interested in.
+
+
[…]
