### 2018-06-17 1fe9169
shorten README
