+++ README.Rmd
@@ -45,7 +45,7 @@ install.packages("devtools")
 devtools::install_github("SCAR/antanym")
 ```
 
-## Usage
+## Example usage
 
 Start by fetching the names data from the host server. Here we use a temporary cache so that we can re-load it later in the session without needing to re-download it:
 
@@ -64,15 +64,9 @@ Corresponding to how many distinct features?
 length(unique(g$scar_common_id))
 ```
 
-Find all names associated with feature 1589 (Booth Island) and the country of origin of each name:
+Find names starting with "Slom":
 ```{r message = FALSE, warning = FALSE}
-an_filter(g, feature_ids = 1589)[, c("place_name", "origin")]
-```
-
-Choose one name per feature, preferring the Polish name where there is one:
-```{r message = FALSE, warning = FALSE}
-g <- an_preferred(g, origin = "Poland")
-nrow(g)
+an_filter(g, query = "^Slom")[, c("place_name", "longitude", "latitude")]
 ```
 
 Find islands within 20km of 100E, 66S:
@@ -85,60 +79,28 @@ nms <- g %>% an_filter(feature_type = "Island") %>% an_near(loc = c(100, -66), m
 nms[, c("place_name", "longitude", "latitude")]
 ```
 
-Find names starting with "Slom":
+## Resolving multiple names per feature
+
+As noted above, the CGA is a composite gazetteer and so there may be multiple names associated with a given feature. For example, we can see all names associated with feature 1589 (Booth Island) and the country of origin of each name:
+
 ```{r message = FALSE, warning = FALSE}
-an_filter(g, query = "^Slom")[, c("place_name", "longitude", "latitude")]
+an_filter(g, feature_ids = 1589)[, c("place_name", "origin")]
 ```
 
-### Name suggestions
+The `an_preferred` function can help with finding one name per feature. It takes an `origin` parameter that specifies one or more preferred name issuers (countries or organisations). For features that have multiple names (e.g. have been named by multiple countries) a single name will be chosen, preferring names from the specified \code{origin} bodies where possible.
+
+We start with `r nrow(g)` names in the full CGA, corresponding to `r length(unique(g$scar_common_id))` distinct features. Choose one name per feature, preferring the Polish name where there is one, and the German name as a second preference:
+```{r}
+g <- an_preferred(g, origin = c("Poland", "Germany"))
+```
+
+Now we have `r nrow(g)` names in our data frame, corresponding to the same `r length(unique(g$scar_common_id))` distinct features.
+
+## Name suggestions
 
 Antanym includes an experimental function that will suggest which features might be best to add names to on a given map. These suggestions are based on maps prepared by expert cartographers, and the features that were explicitly named on those maps.
 
-Let's say we are preparing a figure of the greater Prydz Bay region (60&ndash;90 &deg;E, 65&ndash;70 &deg;S), to be shown at 80mm x 80mm in size (this is approximately a 1:10M scale map). Let's plot all of the place names in this region:
-
-```{r map0, message = FALSE, warning = FALSE, dpi = 72, fig.width = 9, fig.height = 6, fig.align = "center"}
-my_longitude <- c(60, 90)
-my_latitude <- c(-70, -65)
-
-this_names <- an_filter(g, extent = c(my_longitude, my_latitude))
-
-library(rworldmap)
-map <- getMap(resolution = "low")
-plot(map, xlim = my_longitude + c(-7, 4), ylim = my_latitude, col = "grey50") ## extra xlim space for labels
-points(this_names$longitude, this_names$latitude, pch = 21, bg = "green", cex = 2)
-pos <- rep(c(1, 2, 3, 4), ceiling(nrow(this_names)/4)) ## alternate positions of labels to reduce overlap
-pos[order(this_names$longitude)] <- pos[1:nrow(this_names)]
[…]
