+++ README.md
@@ -10,7 +10,7 @@ Overview
 
 This R package provides easy access to Antarctic geographic place name information, and tools for working with those names.
 
-The authoritative source of place names in Antarctica is the Composite Gazetteer of Antarctica (CGA), which is produced by the Scientific Committee on Antarctic Research (SCAR). The CGA consists of approximately 37,000 names corresponding to 19,000 distinct features. It covers features south of 60 &deg;S, including terrestrial and undersea or under-ice.
+The authoritative source of place names in Antarctica is the Composite Gazetteer of Antarctica (CGA), which is produced by the Scientific Committee on Antarctic Research (SCAR). The CGA consists of approximately 37,000 names corresponding to 19,000 distinct features. It covers features south of 60 °S, including terrestrial and undersea or under-ice.
 
 There is no single naming authority responsible for place names in Antarctica because it does not fall under the sovereignty of any one nation. In general, individual countries have administrative bodies that are responsible for their national policy on, and authorisation and use of, Antarctic names. The CGA is a compilation of place names that have been submitted by representatives of national names committees from 22 countries.
 
@@ -32,8 +32,8 @@ install.packages("devtools")
 devtools::install_github("SCAR/antanym")
 ```
 
-Usage
------
+Example usage
+-------------
 
 Start by fetching the names data from the host server. Here we use a temporary cache so that we can re-load it later in the session without needing to re-download it:
 
@@ -56,12 +56,44 @@ length(unique(g$scar_common_id))
 #> [1] 19569
 ```
 
-Find all names associated with feature 1589 (Booth Island) and the country of origin of each name:
+Find names starting with "Slom":
+
+``` r
+an_filter(g, query = "^Slom")[, c("place_name", "longitude", "latitude")]
+#> # A tibble: 3 x 3
+#>   place_name     longitude latitude
+#>   <chr>              <dbl>    <dbl>
+#> 1 Sloman Glacier     -68.6    -67.7
+#> 2 Sloman Glacier     -68.6    -67.7
+#> 3 Slomer Cove        -59.4    -63.8
+```
+
+Find islands within 20km of 100E, 66S:
+
+``` r
+nms <- an_near(an_filter(g, feature_type = "Island"), loc = c(100, -66), max_distance = 20)
+
+## or equivalently, using the pipe operator
+nms <- g %>% an_filter(feature_type = "Island") %>% an_near(loc = c(100, -66), max_distance = 20)
+
+nms[, c("place_name", "longitude", "latitude")]
+#> # A tibble: 3 x 3
+#>   place_name    longitude latitude
+#>   <chr>             <dbl>    <dbl>
+#> 1 Foster Island       100    -66.1
+#> 2 Severnyj holm       100    -66.1
+#> 3 Foster Island       100    -66.1
+```
+
+Resolving multiple names per feature
+------------------------------------
+
+As noted above, the CGA is a composite gazetteer and so there may be multiple names associated with a given feature. For example, we can see all names associated with feature 1589 (Booth Island) and the country of origin of each name:
 
 ``` r
 an_filter(g, feature_ids = 1589)[, c("place_name", "origin")]
 #> # A tibble: 7 x 2
-#>   place_name   country_name            
+#>   place_name   origin                  
 #>   <chr>        <chr>                   
 #> 1 Booth, isla  Argentina               
 #> 2 Wandel, Ile  Belgium                 
@@ -72,97 +104,22 @@ an_filter(g, feature_ids = 1589)[, c("place_name", "origin")]
 #> 7 Booth Island United States of America
 ```
 
-Choose one name per feature, preferring the Polish name where there is one:
[…]
