+++ NAMESPACE
@@ -46,6 +46,7 @@ importFrom(sp,spTransform)
 importFrom(stats,dist)
 importFrom(stats,na.omit)
 importFrom(stats,predict)
+importFrom(stats,setNames)
 importFrom(stringi,stri_trans_general)
 importFrom(utils,download.file)
 importFrom(utils,write.csv)
