### 2018-06-17 0e7f0ba
minor doc revision
