### 2018-06-17 46b2d5c
sp:: namespace on SpatialPoints usage
