### 2018-06-17 7882866
data structure section minor rev
