### 2018-06-17 b9b0797
revise an_suggest doc, add test
