+++ R/suggest.R
@@ -5,7 +5,7 @@
 #'
 #' @param gaz data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}
 #' @param map_scale numeric: the scale of the map (e.g. 20e6 for a 1:20M map). If \code{map_scale} is not provided, it will be estimated from \code{map_extent} and \code{map_dimensions}
-#' @param map_extent raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. Not required if \code{map_scale} is provided
+#' @param map_extent raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. This is required if \code{map_scale} is not provided, and optional if \code{map_scale} is provided (if \code{map_extent} is provided in this situation then the \code{gaz} data frame will be filtered to this extent before the suggestion algorithm is applied; otherwise all names in \code{gaz} will be considered)
 #' @param map_dimensions numeric: 2-element numeric giving width and height of the map, in mm. Not required if \code{map_scale} is provided
 #'
 #' @return data.frame of names with a "score" column added. Score values range from 0 to 1. The data frame will be sorted in descending score order. Names with higher scores are those that are suggested as the most suitable for display.
@@ -22,9 +22,11 @@
 #'
 #'  ## suggested names for a 100x100 mm map covering 60-90E, 70-60S
 #'  ##  (this is about a 1:12M scale map)
-#'
 #'  suggested <- an_suggest(g, map_extent = c(60, 90, -70, -60), map_dimensions = c(100, 100))
 #'  head(suggested, 20) ## top 20 names
+#'
+#'  ## an equivalent result can be achieved by supplying map scale and extent
+#'  suggested <- an_suggest(g, map_scale = 12e6, map_extent = c(60, 90, -70, -60))
 #' }
 #' @export
 an_suggest <- function(gaz, map_scale, map_extent, map_dimensions) {
@@ -36,7 +38,7 @@ an_suggest <- function(gaz, map_scale, map_extent, map_dimensions) {
         if (missing(map_extent) || missing(map_dimensions)) stop("need either map_scale, or map_dimensions and map_extent")
         map_scale <- an_mapscale(map_dimensions, map_extent)
     }
-    temp <- an_filter(gaz, extent = map_extent)
+    if (!missing(map_extent)) gaz <- an_filter(gaz, extent = map_extent)
     if (map_scale>=10e6) {
         ## for scale >= 10e6: we have full coverage (nearly so for 12mill) of all scar_common_ids,
         ## so we use per-feature predictions. That is, for every feature we have a regression tree
