+++ man/an_thin.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/thin.R
 \name{an_thin}
 \alias{an_thin}
 \title{Thin names to give approximately uniform spatial coverage}
