+++ man/an_suggest.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/suggest.R
 \name{an_suggest}
 \alias{an_suggest}
 \title{Suggest names for a map (experimental)}
@@ -11,7 +11,7 @@ an_suggest(gaz, map_scale, map_extent, map_dimensions)
 
 \item{map_scale}{numeric: the scale of the map (e.g. 20e6 for a 1:20M map). If \code{map_scale} is not provided, it will be estimated from \code{map_extent} and \code{map_dimensions}}
 
-\item{map_extent}{raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. Not required if \code{map_scale} is provided}
+\item{map_extent}{raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. This is required if \code{map_scale} is not provided, and optional if \code{map_scale} is provided (if \code{map_extent} is provided in this situation then the \code{gaz} data frame will be filtered to this extent before the suggestion algorithm is applied; otherwise all names in \code{gaz} will be considered)}
 
 \item{map_dimensions}{numeric: 2-element numeric giving width and height of the map, in mm. Not required if \code{map_scale} is provided}
 }
@@ -32,9 +32,11 @@ This is an experimental function and currently only implemented for \code{map_sc
 
  ## suggested names for a 100x100 mm map covering 60-90E, 70-60S
  ##  (this is about a 1:12M scale map)
-
  suggested <- an_suggest(g, map_extent = c(60, 90, -70, -60), map_dimensions = c(100, 100))
  head(suggested, 20) ## top 20 names
+
+ ## an equivalent result can be achieved by supplying map scale and extent
+ suggested <- an_suggest(g, map_scale = 12e6, map_extent = c(60, 90, -70, -60))
 }
 }
 \seealso{
