### 2018-06-17 56399c1
an_mapscale and an_suggest to take various extent formats
