+++ R/suggest.R
@@ -5,7 +5,7 @@
 #'
 #' @param gaz data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}
 #' @param map_scale numeric: the scale of the map (e.g. 20e6 for a 1:20M map). If \code{map_scale} is not provided, it will be estimated from \code{map_extent} and \code{map_dimensions}
-#' @param map_extent raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. This is required if \code{map_scale} is not provided, and optional if \code{map_scale} is provided (if \code{map_extent} is provided in this situation then the \code{gaz} data frame will be filtered to this extent before the suggestion algorithm is applied; otherwise all names in \code{gaz} will be considered)
+#' @param map_extent vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. This is required if \code{map_scale} is not provided, and optional if \code{map_scale} is provided (if \code{map_extent} is provided in this situation then the \code{gaz} data frame will be filtered to this extent before the suggestion algorithm is applied; otherwise all names in \code{gaz} will be considered). \code{map_extent} can also be passed as a raster Extent object, a Raster object (in which case its extent will be used), a Spatial object (in which case the bounding box of the object will be used as the extent), or a matrix (in which case it will be assumed to be the output of \code{sp::bbox})
 #' @param map_dimensions numeric: 2-element numeric giving width and height of the map, in mm. Not required if \code{map_scale} is provided
 #'
 #' @return data.frame of names with a "score" column added. Score values range from 0 to 1. The data frame will be sorted in descending score order. Names with higher scores are those that are suggested as the most suitable for display.
@@ -32,8 +32,7 @@
 an_suggest <- function(gaz, map_scale, map_extent, map_dimensions) {
     assert_that(inherits(gaz, c("data.frame", "SpatialPointsDataFrame")))
     if (!missing(map_extent))
-        assert_that((is.numeric(map_extent) && length(map_extent) == 4) || inherits(map_extent, "Extent"))
-
+        assert_that((is.numeric(map_extent) && length(map_extent) == 4) || inherits(map_extent, "Extent") || inherits(map_extent, "Raster") || inherits(map_extent, "Spatial"))
     if (missing(map_scale)) {
         if (missing(map_extent) || missing(map_dimensions)) stop("need either map_scale, or map_dimensions and map_extent")
         map_scale <- an_mapscale(map_dimensions, map_extent)
@@ -44,15 +43,15 @@ an_suggest <- function(gaz, map_scale, map_extent, map_dimensions) {
         ## so we use per-feature predictions. That is, for every feature we have a regression tree
         ## that models the probability of this feature being shown on a map of given scale
         ## the uid_fits object lives in sysdata.rda
-        temp$score <- 0
-        temp$scale <- map_scale
-        idx <- which(temp$scar_common_id %in% uid)
+        gaz$score <- 0
+        gaz$scale <- map_scale
+        idx <- which(gaz$scar_common_id %in% uid)
         for (i in idx) {
-            fidx <- which(temp$scar_common_id[i] == uid)
+            fidx <- which(gaz$scar_common_id[i] == uid)
             if (inherits(uid_fits[[fidx]], "C5.0")) {
-                temp$score[i] <- predict(uid_fits[[fidx]], newdata = temp[i,], type = "prob")[, 2]
[…]
