### 2018-06-17 193ac3f
expand description of suggestion alg
