### 2018-06-17 d1f287c
minor revs
