### 2018-06-17 b8580cc
minor version bump, contributors updated
