+++ DESCRIPTION
@@ -1,11 +1,13 @@
 Package: antanym
 Type: Package
 Title: Antarctic Geographic Place Names
-Version: 0.3.1.9000
+Version: 0.4.0
 Authors@R: c(person("Ben", "Raymond", email = "ben.raymond@aad.gov.au", role = c("aut", "cre")),
 	     person("Michael", "Sumner", email = "michael.sumner@aad.gov.au", role="aut"),
-	     person("Ursula", "Harris", role = "ctb"),
+	     person("John", "Baumgartner", role = c("ctb", "rev")),
+	     person("Lorenzo", "Busetto", role = c("ctb", "rev")),
 	     person("Andrew", "Cowie", role = "ctb"),
+	     person("Ursula", "Harris", role = "ctb"),
 	     person("Fraser", "Morgan", role = "ctb")
 	    )
 Description: Antarctic geographic names from the Composite Gazetteer of Antarctica, and functions for working with those place names.
