### 2018-06-17 55ba1f7
tryCatch around download, with retry
