### 2018-06-17 e04a2bb
formatting, add recent changes to readme
