+++ man/an_near.Rd
@@ -38,7 +38,7 @@ Find placenames near a given location
 
 }
 \references{
-\url{https://www.scar.org/data-products/place-names/}
+\url{https://data.aad.gov.au/aadc/gaz/scar/}, \url{https://www.scar.org/data-products/place-names/}
 }
 \seealso{
 \code{\link{an_read}}
