+++ man/an_preferred.Rd
@@ -35,7 +35,7 @@ The Composite Gazetteer of Antarctica is a compilation of place names provided b
 
 }
 \references{
-\url{https://www.scar.org/data-products/place-names/}
+\url{https://data.aad.gov.au/aadc/gaz/scar/}, \url{https://www.scar.org/data-products/place-names/}
 }
 \seealso{
 \code{\link{an_read}}, \code{\link{an_origins}}
