### 2018-06-18 c619ad2
don't rely on content(as = 'parsed')
