### 2018-06-19 474f176
cran badges [ci skip]
