### 2018-06-20 #198 @sckott
@lbusett @johnbaums are you happy with revisions/comments from the maintainer?  feel free to make any suggestions/comments here if you have anything further. if nothing, that's okay too, either way let me know
