### 2018-06-20 #198 @raymondben
(Comments particularly welcome on the revised `an_filter(query=...)` behaviour - is it more intuitive now?)
