### 2018-07-09 #198 @lbusett
@raymondben @sckott

 sorry for being late on this, but I had very limited time in the last two weeks.

I had a look at the changes: great work! All points raised in my review were nicely addressed. I particularly like the changes introduced in the `an_filter()` function and the improvements in the vignette (one small thing: maybe you could specify in the vignette that when using session = "persistent" the dataset is saved in rappdirs::user_cache_dir(), as it is done in the documentation).  The package is now ready for approval, in my opinion.

Lorenzo
