### 2018-07-09 #198 @sckott
thanks @lbusett !

@johnbaums are you happy with changes/responses?
