### 2018-07-10 a904ca5
mention rappdirs in vignette for determining persistent cache location
