### 2018-07-13 #198 @johnbaums
@sckott @raymondben

Thanks for looping me in on the changes, and sorry for my abysmally slow handling of GitHub notifications. I'm very satisfied - great package!
