### 2018-07-13 e08f303
change SCAR github refs to ropensci ready for repo transfer [ci skip]
