+++ vignettes/antanym.Rmd
@@ -36,8 +36,8 @@ The SCAR Composite Gazetteer of Antarctica is made available under a CC-BY licen
 
 
 ```{r eval = FALSE}
-install.packages("devtools")
-devtools::install_github("SCAR/antanym")
+install.packages("remotes")
+remotes::install_github("ropensci/antanym")
 ```
 
 ## Usage
