+++ DESCRIPTION
@@ -11,8 +11,8 @@ Authors@R: c(person("Ben", "Raymond", email = "ben.raymond@aad.gov.au", role = c
 	     person("Fraser", "Morgan", role = "ctb")
 	    )
 Description: Antarctic geographic names from the Composite Gazetteer of Antarctica, and functions for working with those place names.
-URL: https://github.com/SCAR/antanym
-BugReports: https://github.com/SCAR/antanym/issues
+URL: https://github.com/ropensci/antanym
+BugReports: https://github.com/ropensci/antanym/issues
 Depends:
     R (>= 3.3.0)
 License: MIT + file LICENSE
@@ -42,4 +42,4 @@ RoxygenNote: 6.0.1
 VignetteBuilder: knitr
 X-schema.org-applicationCategory: Antarctic/Southern Ocean
 X-schema.org-keywords: Antarctic, Southern Ocean, place names, gazetteer
-X-schema.org-isPartOf: https://scar.org
+X-schema.org-isPartOf:  https://ropensci.org, https://scar.org
