### 2018-07-14 4f9bdc0
add ropensci badge
