### 2018-07-15 #198 @sckott
See comment above about a blog post. thoughts on that?
