### 2018-07-16 #198 @raymondben
Yup, is a good idea. We're thinking of a more comprehensive post (this package plus others relevant to Antarctic stuff), but will be aiming to do *something*, anyway.
